@@ -1,7 +1,7 @@
 import { promisify } from 'util';
 import winston from 'winston';
-const child_process = require('child_process');
-const exec = promisify(child_process.exec);
+import { exec } from 'child_process';
+const execPromise = promisify(exec);
 
 /**
  * Sends ssh commands to the host system
@@ -13,11 +13,36 @@ export default class SshService {
   /**
    * Sends ssh command to host
    */
-  static async sendCommand(command: string, needSudo = false, env?: string[]): Promise<any> {
+  static async sendCommand(
+    command: string,
+    needSudo = false,
+    env: string[] = []
+  ): Promise<{ stdout: string | Buffer; stderr: string | Buffer }> {
     const logPrefix = `SshService::sendCommand`;
     const envVariables = env?.join(' ');
-    winston.debug(`${logPrefix} received command ${needSudo ? 'with sudo' : ''}: ${command} ${env ? `with env: ${envVariables}` : ''}`);
+    const chainedCommand = `${needSudo ? 'sudo ' : ''}${
+      env.length > 0 ? `${envVariables} ` : ''
+    }${command}`.replace(/"/g, `\"`);
+    winston.debug(`${logPrefix} sending: "${chainedCommand}"`);
 
-    return exec(`${this.sshCommand}${needSudo ? ' sudo ' : ' '}${envVariables || ' '} ${command}`);
+    return (
+      execPromise(`${this.sshCommand} "${chainedCommand}"`)
+        // Log exec return without changing error and response behavior
+        .then(({ stderr, stdout }) => {
+          const logMsg = `${logPrefix} received from command: "${chainedCommand}" \n
+stdout: ${stdout}
+stderr: ${stderr}`;
+          winston.debug(logMsg);
+          return Promise.resolve({ stderr, stdout });
+        })
+        .catch((err) => {
+          const logMsg = `${logPrefix} catch error from command: "${chainedCommand}"\n
+error: ${JSON.stringify(err)}
+stdout: ${err.stdout}
+stderr: ${err.stderr}`;
+          winston.error(logMsg);
+          return Promise.reject(err);
+        })
+    );
   }
 }
