@@ -1,5 +1,5 @@
 var replace = require('replace-in-file');
-var buildVersion = process.argv[2];
+var buildVersion = process.argv[2] || 'unknown';
 const options = {
   files: 'src/environments/environment.prod.ts',
   from: /{BUILD_VERSION}/g,
@@ -8,7 +8,7 @@ const options = {
 };
 
 try {
-  let changedFiles = replace.sync(options);
+  replace.sync(options);
   console.log('Build version set: ' + buildVersion);
 } catch (error) {
   console.error('Error occurred:', error);
