@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
-import { Observable, of } from 'rxjs';
+import { Observable, of, Subscription } from 'rxjs';
 import { filter, map, mergeMap, shareReplay } from 'rxjs/operators';
 import { ActivatedRoute, NavigationEnd, Router } from '@angular/router';
 import { AuthService } from '../../shared';
@@ -18,30 +18,42 @@ export class LayoutComponent {
       shareReplay()
     );
 
-  routeData$ = this.router.events.pipe(
-    filter((event) => event instanceof NavigationEnd),
-    map(() => this.route),
-    mergeMap((route) => {
-      if (route.firstChild) {
-        return route.firstChild!.data;
-      }
-
-      return of({});
-    })
-  );
+  noLayout = false;
+  resetPasswordLayout = false;
 
   get supportHref() {
     return `${window.location.protocol}//${window.location.hostname}/help/docs/`;
   }
+
+  subs = new Subscription();
 
   constructor(
     private breakpointObserver: BreakpointObserver,
     private router: Router,
     private route: ActivatedRoute,
     private auth: AuthService
-  ) {}
+  ) {
+    this.subs.add(this.router.events.pipe(
+      filter((event) => event instanceof NavigationEnd),
+      map(() => this.route),
+      mergeMap((route) => {
+        if (route.firstChild) {
+          return route.firstChild!.data;
+        }
+  
+        return of({});
+      })
+    ).subscribe(({resetPasswordLayout, noLayout}) => {
+      this.resetPasswordLayout = resetPasswordLayout;
+      this.noLayout = noLayout;
+    }));
+  }
 
   async logout() {
     await this.auth.logout();
+  }
+
+  ngOnDestroy() {
+    this.subs.unsubscribe();
   }
 }
