@@ -11,9 +11,9 @@ const routes: Routes = [
     component: ChangePasswordComponent,
   },
   {
-    path: 'forced-change-password',
+    path: 'reset-password',
     component: ChangePasswordComponent,
-    data: { noLayout: true },
+    data: { resetPasswordLayout: true },
   },
 ];
 
