@@ -1,12 +1,17 @@
 import winston from 'winston';
 import { Logger } from './modules/Logger';
 import { BootstrapManager } from './modules/BootstrapManager';
+import { System } from './modules/System';
 
 Logger.init();
 
-winston.info('MDC light starting...');
+winston.error('MDC Flex starting...');
+winston.error(
+  `MDC Flex runtime version: ${process.env.MDC_LIGHT_RUNTIME_VERSION}`
+);
+winston.error(`MDC Flex OS version: ${new System().readOsVersion()}`);
 
 const bootstrapManager = new BootstrapManager();
 bootstrapManager.start();
 
-winston.info('MDC light started');
+winston.error('MDC Flex started');
