@@ -59,8 +59,8 @@ export class MTConnectDataSink extends DataSink {
     const logPrefix = `${this.name}::init`;
     winston.info(`${logPrefix} initializing.`);
 
-    if(!this.licenseChecker.isLicensed) {
-      winston.warn(`${logPrefix} no valid license found. Stop initializing.`)
+    if (!this.isLicensed) {
+      winston.warn(`${logPrefix} no valid license found. Stop initializing.`);
       this.updateCurrentStatus(LifecycleEventStatus.NoLicense);
       return this;
     }
