@@ -37,7 +37,7 @@ export class OPCUADataSink extends DataSink {
     this.opcuaAdapter = new OPCUAAdapter({
       config: options.dataSinkConfig,
       generalConfig: options.generalConfig,
-      runtimeConfig: options.runtimeConfig,
+      runtimeConfig: options.runtimeConfig
     });
     this.generalConfig = options.generalConfig;
   }
@@ -46,8 +46,8 @@ export class OPCUADataSink extends DataSink {
     const logPrefix = `${this.name}::init`;
     winston.info(`${logPrefix} initializing.`);
 
-    if(!this.licenseChecker.isLicensed) {
-      winston.warn(`${logPrefix} no valid license found. Stop initializing.`)
+    if (!this.isLicensed) {
+      winston.warn(`${logPrefix} no valid license found. Stop initializing.`);
       this.updateCurrentStatus(LifecycleEventStatus.NoLicense);
       return this;
     }
