@@ -3,12 +3,28 @@ import { ConfigManager } from '../ConfigManager';
 import { IVirtualDataPointConfig } from '../ConfigManager/interfaces';
 import { CounterManager } from '../CounterManager';
 import { DataPointCache } from '../DatapointCache';
-import { IDataSourceMeasurementEvent } from '../Southbound/DataSources/interfaces';
+import {
+  IDataSourceMeasurementEvent,
+  IMeasurement
+} from '../Southbound/DataSources/interfaces';
+import { SynchronousIntervalScheduler } from '../SyncScheduler';
 
 interface IVirtualDataPointManagerParams {
   configManager: ConfigManager;
   cache: DataPointCache;
 }
+
+type LogEntry = {
+  count: number;
+  log: string;
+};
+
+type LogSummary = {
+  error: LogEntry[];
+  warn: LogEntry[];
+  info: LogEntry[];
+  debug: LogEntry[];
+};
 
 /**
  * Calculates virtual datapoints
@@ -18,21 +34,83 @@ export class VirtualDataPointManager {
   private config: IVirtualDataPointConfig[] = null;
   private cache: DataPointCache;
   private counters: CounterManager;
+  private scheduler: SynchronousIntervalScheduler;
+  private logSchedulerListenerId: number;
+  private logSummary: LogSummary = {
+    error: [],
+    warn: [],
+    info: [],
+    debug: []
+  };
 
   private static className: string = VirtualDataPointManager.name;
 
   constructor(params: IVirtualDataPointManagerParams) {
-    params.configManager.once('configsLoaded', () => {
-      return this.init();
-    });
+    params.configManager.once('configsLoaded', () => this.init());
+    params.configManager.on('configChange', () => this.updateConfig());
 
     this.configManager = params.configManager;
     this.cache = params.cache;
     this.counters = new CounterManager(this.configManager);
+    this.scheduler = SynchronousIntervalScheduler.getInstance();
   }
 
   private init() {
+    this.updateConfig();
+    this.setupLogCycle();
+  }
+
+  private updateConfig() {
+    const logPrefix = `${VirtualDataPointManager.className}::updateConfig`;
+    winston.debug(`${logPrefix} refreshing config`);
+
     this.config = this.configManager.config.virtualDataPoints;
+  }
+
+  /**
+   * Setup log cycle for summary logging
+   */
+  protected setupLogCycle(defaultFrequency: number = 15 * 60 * 1000): void {
+    if (this.logSchedulerListenerId) return;
+
+    const logPrefix = `${VirtualDataPointManager.className}::setupLogCycle`;
+    winston.debug(`${logPrefix} setup log cycle`);
+
+    this.logSchedulerListenerId = this.scheduler.addListener(
+      [defaultFrequency],
+      this.printSummaryLogs.bind(this)
+    );
+  }
+
+  /**
+   * Logs summary of logs to not spam logs every data point ready cycle
+   */
+  protected printSummaryLogs(): void {
+    const logPrefix = `${VirtualDataPointManager.className}::logSummary`;
+
+    Object.keys(this.logSummary).forEach((level: keyof LogSummary) => {
+      if (this.logSummary[level].length > 0) {
+        for (const entry of this.logSummary[level]) {
+          winston[level](
+            `${logPrefix} Count: ${entry.count} Log: ${entry.log}`
+          );
+        }
+        this.logSummary[level] = [];
+      }
+    });
+  }
+
+  /**
+   * Adding log to summary. If Same log already exits, counting up
+   */
+  protected addSummaryLog(level: keyof LogSummary, log: string): void {
+    const entry = this.logSummary[level].find((entry) => entry.log === log);
+
+    if (entry) {
+      entry.count = entry.count + 1;
+    } else {
+      this.logSummary[level].push({ log, count: 1 });
+    }
   }
 
   /**
@@ -51,6 +129,45 @@ export class VirtualDataPointManager {
   }
 
   /**
+   * Calculates the sum of a  list of virtual data points.
+   *
+   * @param  {IDataSourceMeasurementEvent[]} sourceEvents
+   * @param  {IVirtualDataPointConfig} config
+   * @returns boolean
+   */
+  private sum(
+    sourceEvents: IDataSourceMeasurementEvent[],
+    config: IVirtualDataPointConfig
+  ): number | null {
+    const logPrefix = `VirtualDataPointManager::sum`;
+    if (config.operationType !== 'sum') {
+      winston.error(`${logPrefix} called with invalid operation type.`);
+      return null;
+    }
+    if (sourceEvents.length < 2) {
+      winston.warn(
+        `${logPrefix} virtual data point (${config.id}) requires at least 2 sources!`
+      );
+      return null;
+    }
+    try {
+      return sourceEvents.reduce(
+        (previousValue, { measurement: { value } }) => {
+          if (typeof value !== 'number') {
+            // Not addable
+            throw Error(`${value} is not a number.`);
+          }
+          return value + previousValue;
+        },
+        0
+      );
+    } catch (err) {
+      winston.error(`${logPrefix} ${err.message}`);
+      return null;
+    }
+  }
+
+  /**
    * Calculates virtual data points from type and
    *
    * @param  {IDataSourceMeasurementEvent[]} sourceEvents
@@ -62,7 +179,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): boolean | null {
     if (sourceEvents.length < 2) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires at least 2 sources!`
       );
       return null;
@@ -93,7 +211,8 @@ export class VirtualDataPointManager {
       equal ? '(Equal)' : ''
     }`;
     if (sourceEvents.length > 1 || sourceEvents.length === 0) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `${logPrefix} is only available for one source but receive ${sourceEvents.length}`
       );
       return null;
@@ -102,7 +221,8 @@ export class VirtualDataPointManager {
     // @ts-ignore
     const compare = parseFloat(config.comparativeValue);
     if (Number.isNaN(compare)) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `${logPrefix} got ${compare} for comparison but it's not a number`
       );
       return null;
@@ -121,7 +241,7 @@ export class VirtualDataPointManager {
       case 'string': {
         const parsed = parseFloat(value);
         if (isNaN(parsed)) {
-          winston.error(`${logPrefix} no valid number.`);
+          this.addSummaryLog('error', `${logPrefix} no valid number.`);
           return null;
         }
         if (equal && Math.abs(compare - parsed) < 1e-9) {
@@ -130,11 +250,17 @@ export class VirtualDataPointManager {
         return parsed > compare;
       }
       case 'boolean': {
-        winston.warn(`${logPrefix} boolean is not a valid compare value.`);
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} boolean is not a valid compare value.`
+        );
         return null;
       }
       default: {
-        winston.warn(`${logPrefix} invalid value type: ${typeof value}`);
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} invalid value type: ${typeof value}`
+        );
         return null;
       }
     }
@@ -157,7 +283,8 @@ export class VirtualDataPointManager {
       equal ? '(Equal)' : ''
     }`;
     if (sourceEvents.length > 1 || sourceEvents.length === 0) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `${logPrefix} is only available for one source but receive ${sourceEvents.length}`
       );
       return null;
@@ -166,7 +293,8 @@ export class VirtualDataPointManager {
     // @ts-ignore
     const compare = parseFloat(config.comparativeValue);
     if (Number.isNaN(compare)) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `${logPrefix} got ${compare} for comparison but it's not a number`
       );
       return null;
@@ -186,7 +314,7 @@ export class VirtualDataPointManager {
       case 'string': {
         const parsed = parseFloat(value);
         if (isNaN(parsed)) {
-          winston.error(`${logPrefix} no valid number.`);
+          this.addSummaryLog('error', `${logPrefix} no valid number.`);
           return null;
         }
         if (equal && Math.abs(compare - parsed) < 1e-9) {
@@ -195,11 +323,17 @@ export class VirtualDataPointManager {
         return parsed < compare;
       }
       case 'boolean': {
-        winston.warn(`${logPrefix} boolean is not a valid compare value.`);
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} boolean is not a valid compare value.`
+        );
         return null;
       }
       default: {
-        winston.warn(`${logPrefix} invalid value type: ${typeof value}`);
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} invalid value type: ${typeof value}`
+        );
         return null;
       }
     }
@@ -218,7 +352,8 @@ export class VirtualDataPointManager {
   ): boolean | null {
     const logPrefix = `${this.constructor.name}::equal`;
     if (sourceEvents.length > 1 || sourceEvents.length === 0) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `${logPrefix} is only available for one source but receive ${sourceEvents.length}`
       );
       return null;
@@ -243,7 +378,8 @@ export class VirtualDataPointManager {
       }
       case 'string': {
         if (typeof compare !== 'string') {
-          winston.warn(
+          this.addSummaryLog(
+            'warn',
             `${logPrefix} try to compare string with non string value: ${compare}`
           );
           return null;
@@ -251,11 +387,17 @@ export class VirtualDataPointManager {
         return value.trim() === compare.trim();
       }
       case 'boolean': {
-        winston.warn(`${logPrefix} boolean is not a valid compare value.`);
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} boolean is not a valid compare value.`
+        );
         return null;
       }
       default: {
-        winston.warn(`${logPrefix} invalid value type: ${typeof value}`);
+        this.addSummaryLog(
+          'warn',
+          `${logPrefix} invalid value type: ${typeof value}`
+        );
         return null;
       }
     }
@@ -287,7 +429,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): boolean | null {
     if (sourceEvents.length < 2) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires at least 2 sources!`
       );
       return null;
@@ -313,7 +456,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): boolean | null {
     if (sourceEvents.length !== 1) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires exactly 1 source!`
       );
       return null;
@@ -334,7 +478,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): number | null {
     if (sourceEvents.length !== 1) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires exactly 1 source!`
       );
       return null;
@@ -352,6 +497,56 @@ export class VirtualDataPointManager {
   }
 
   /**
+   * Check 'enumerated' grouped virtual data point and return the 'returnValueIfTrue' of the highest true value.
+   *
+   * @param sourceEvents
+   * @param config
+   * @returns string
+   */
+  private enumeration(
+    sourceEvents: IDataSourceMeasurementEvent[],
+    config: IVirtualDataPointConfig
+  ): string | null {
+    const logPrefix = `${this.constructor.name}::enumeration`;
+    if (config.operationType !== 'enumeration') {
+      this.addSummaryLog(
+        'error',
+        `${logPrefix} receive invalid operation type: ${config.operationType}`
+      );
+      return null;
+    }
+    if (typeof config.enumeration === undefined) {
+      this.addSummaryLog(
+        'error',
+        `${logPrefix} no enumeration configuration found`
+      );
+      return null;
+    }
+
+    // Iterate over sorted by high prio array
+    for (const entry of config.enumeration.items.sort((a, b) => {
+      return b.priority - a.priority;
+    })) {
+      const hit = !!sourceEvents.find((event) => {
+        // winston.debug(
+        //   `${logPrefix} searching for entry ${entry.source} found ${event.measurement.id}`
+        // );
+        // winston.debug(
+        //   `${logPrefix} measurement is: ${!!event.measurement.value}`
+        // );
+        return (
+          event.measurement.id === entry.source && !!event.measurement.value
+        );
+      });
+      if (hit) {
+        return entry.returnValueIfTrue;
+      }
+    }
+    // No true value in list
+    return config.enumeration?.defaultValue || null;
+  }
+
+  /**
    * Calculates virtual data points from type thresholds
    *
    * @param  {IDataSourceMeasurementEvent[]} sourceEvents
@@ -363,7 +558,8 @@ export class VirtualDataPointManager {
     config: IVirtualDataPointConfig
   ): number | null {
     if (sourceEvents.length !== 1) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `Virtual data point (${config.id}) requires exactly 1 source!`
       );
       return null;
@@ -396,7 +592,7 @@ export class VirtualDataPointManager {
   private calculateValue(
     sourceEvents: IDataSourceMeasurementEvent[],
     config: IVirtualDataPointConfig
-  ): boolean | number | null {
+  ): IMeasurement['value'] {
     switch (config.operationType) {
       case 'and':
         return this.and(sourceEvents, config);
@@ -408,6 +604,8 @@ export class VirtualDataPointManager {
         return this.count(sourceEvents, config);
       case 'thresholds':
         return this.thresholds(sourceEvents, config);
+      case 'enumeration':
+        return this.enumeration(sourceEvents, config);
       case 'greater':
         return this.greater(sourceEvents, config);
       case 'greaterEqual':
@@ -420,10 +618,13 @@ export class VirtualDataPointManager {
         return this.equal(sourceEvents, config);
       case 'unequal':
         return this.unequal(sourceEvents, config);
+      case 'sum':
+        return this.sum(sourceEvents, config);
       default:
-        winston.warn(
-          `Invalid type (${config.operationType}) provided for virtual data point ${config.id}!`
-        );
+        // TODO Only print this once at startup or config change, if invalid
+        // this.addSummaryLog("warn",
+        //   `Invalid type (${config.operationType}) provided for virtual data point ${config.id}!`
+        // );
         return null;
     }
   }
@@ -440,7 +641,8 @@ export class VirtualDataPointManager {
   ): IDataSourceMeasurementEvent[] {
     const logPrefix = `${VirtualDataPointManager.className}::getVirtualEvents`;
     if (this.config === null) {
-      winston.warn(
+      this.addSummaryLog(
+        'warn',
         `${logPrefix} Config not yet loaded. Skipping virtual event calculation`
       );
       return [];
@@ -466,7 +668,8 @@ export class VirtualDataPointManager {
           event = this.cache.getCurrentEvent(sourceId);
         }
         if (!event) {
-          winston.warn(
+          this.addSummaryLog(
+            'warn',
             `Virtual data point ${vdpConfig.id} could not be calculated, because missing event from data source ${sourceId}`
           );
         }
